refactor(core): extract shared quota services into FoundryGate.Core (#119)
The Functions host needs quota resolution but references Data + Domain only, so it
could not reach `FoundryGate.Api.Services.Quota`. Per the orchestrator ruling, a new
`src/FoundryGate.Core` class library (Data + Domain, zero ASP.NET) now holds the
services more than one host needs:

- `FoundryGate.Core.Configuration.GatewayOptions` / `GatewayTier` (both hosts bind the
  same `Gateway` section), plus the `LogAnalyticsWorkspaceId` /
  `LogAnalyticsWorkspaceResourceId` bindings infra has been setting since #108.
- `FoundryGate.Core.Quota`: `IQuotaResolutionService`/`QuotaResolutionService`,
  `QuotaResolution`/`QuotaPreview`, `GatewayTierMapper`/`GatewayTierAssignment`,
  `IGatewayTierSync`/`NullGatewayTierSync`, and `AddQuotaCore()`.
- New `IQuotaResetService` in Core owns the reset itself; the Api's
  `QuotaAllocationService.ResetAsync` resolves the acting admin and delegates, so a
  button and a timer cannot drift.

`ApimGatewayTierSync` stays in the Api (it composes the APIM key service). Api's
`AddQuotaServices()` still registers `GatewayOptions` off `AppSettings` and picks the
tier-sync implementation, then calls `AddQuotaCore()`.
