+++ FoundryGate.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FoundryGate.Domain", "src\F
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FoundryGate.Data", "src\FoundryGate.Data\FoundryGate.Data.csproj", "{A240B5EF-EAF9-488B-9A29-77F8580190CE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FoundryGate.Core", "src\FoundryGate.Core\FoundryGate.Core.csproj", "{764BC212-0A84-47F7-ACB3-234B81F3924C}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FoundryGate.Api", "src\FoundryGate.Api\FoundryGate.Api.csproj", "{E4BAECB9-97C3-4A92-BAF8-4C35F6108E58}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FoundryGate.Functions", "src\FoundryGate.Functions\FoundryGate.Functions.csproj", "{D50C3FE0-D227-410C-90F7-7A2781A231E7}"
@@ -37,6 +39,10 @@ Global
 		{A240B5EF-EAF9-488B-9A29-77F8580190CE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A240B5EF-EAF9-488B-9A29-77F8580190CE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A240B5EF-EAF9-488B-9A29-77F8580190CE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{764BC212-0A84-47F7-ACB3-234B81F3924C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{764BC212-0A84-47F7-ACB3-234B81F3924C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{764BC212-0A84-47F7-ACB3-234B81F3924C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{764BC212-0A84-47F7-ACB3-234B81F3924C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E4BAECB9-97C3-4A92-BAF8-4C35F6108E58}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E4BAECB9-97C3-4A92-BAF8-4C35F6108E58}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E4BAECB9-97C3-4A92-BAF8-4C35F6108E58}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -72,6 +78,7 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{A45829C5-0471-446A-B407-8E02AB8E89DD} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{A240B5EF-EAF9-488B-9A29-77F8580190CE} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{764BC212-0A84-47F7-ACB3-234B81F3924C} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{E4BAECB9-97C3-4A92-BAF8-4C35F6108E58} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{D50C3FE0-D227-410C-90F7-7A2781A231E7} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{E12D7A63-6A6E-456E-BF5C-E18209484F94} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
