+++ src/FoundryGate.Api/Services/Lifecycle/UserLifecycleService.cs
@@ -5,8 +5,8 @@ using FoundryGate.Api.Services.Audit;
 using FoundryGate.Api.Services.Entra;
 using FoundryGate.Api.Services.Identity;
 using FoundryGate.Api.Services.Keys;
-using FoundryGate.Api.Services.Quota;
 using FoundryGate.Api.Services.Requests;
+using FoundryGate.Core.Quota;
 using FoundryGate.Data;
 using FoundryGate.Data.Audit;
 using FoundryGate.Data.Entities;
