+++ src/FoundryGate.Api/Services/Quota/QuotaAllocationService.cs
@@ -1,11 +1,10 @@
 using System.Linq.Expressions;
-using FoundryGate.Api.Services.Audit;
 using FoundryGate.Api.Services.Identity;
+using FoundryGate.Core.Quota;
 using FoundryGate.Data;
 using FoundryGate.Data.Entities;
 using FoundryGate.Data.Extensions;
 using FoundryGate.Domain.Common;
-using FoundryGate.Domain.Constants;
 using FoundryGate.Domain.Quota;
 using FoundryGate.Domain.Quota.Contracts;
 using Microsoft.EntityFrameworkCore;
@@ -14,15 +13,17 @@ namespace FoundryGate.Api.Services.Quota;
 
 /// <summary>
 /// Default <see cref="IQuotaAllocationService"/>: projection-to-record reads over
-/// <see cref="AppDbContext.QuotaAllocations"/>, and the two writers (<c>/me</c> auto-create, manual reset)
-/// that orchestrate <see cref="IQuotaResolutionService"/> and own the <c>SaveChangesAsync</c>.
+/// <see cref="AppDbContext.QuotaAllocations"/>, the <c>/me</c> auto-create that orchestrates
+/// <see cref="IQuotaResolutionService"/> and owns its <c>SaveChangesAsync</c>, and the manual reset,
+/// which resolves the acting admin and hands the work to Core's <see cref="IQuotaResetService"/> —
+/// the same implementation the scheduled Function runs (#38/#119).
 /// </summary>
 public sealed class QuotaAllocationService(
     AppDbContext dbContext,
     IQuotaResolutionService quotaResolution,
+    IQuotaResetService quotaReset,
     GatewayTierMapper tierMapper,
     ICurrentUserAccessor currentUser,
-    IAuditService audit,
     TimeProvider timeProvider,
     ILogger<QuotaAllocationService> logger) : IQuotaAllocationService
 {
@@ -157,123 +158,15 @@ public sealed class QuotaAllocationService(
     /// <inheritdoc />
     public async Task<QuotaResetResult> ResetAsync(CancellationToken cancellationToken)
     {
-        var now = timeProvider.GetUtcNow();
-        var period = BillingPeriod.FromInstant(now);
+        // Resolve the actor before the first gateway call, so a caller without a User row is a 403 that
+        // never moved an APIM subscription (CONVENTIONS.md: every refusal before the external call). The
+        // reset itself lives in Core, shared with the scheduled Function (#38/#119) — the only difference
+        // between a button and a timer is who the audit row names.
+        var actor = await currentUser.GetRequiredUserAsync(cancellationToken);
 
-        var activeUserIds = await dbContext.Users
-            .Where(u => u.IsActive)
-            .OrderBy(u => u.UserId)
-            .Select(u => u.UserId)
-            .ToListAsync(cancellationToken);
+        var outcome = await quotaReset.ResetAsync(QuotaResetTrigger.Admin(actor.UserId), cancellationToken);
 
-        var resolutions = await quotaResolution.ResolveManyAsync(activeUserIds, period, cancellationToken);
-        var touched = resolutions.Select(r => r.Allocation).ToList();
-
-        // Past the commit point once any subscription has been re-scoped at the gateway. A mid-loop ARM
-        // failure aborts before this line and saves nothing (the tier sync adds its audit row without
-        // saving, #156 review), so the reset is all-or-nothing exactly as this method's docs promise.
-        var tierSyncCount = resolutions.Count(r => r.TierSyncRequested);
-        var completionToken = tierSyncCount > 0 ? CancellationToken.None : cancellationToken;
-
-        foreach (var allocation in touched)
-        {
-            Stamp(allocation, now);
-        }
-
-        // One row per run, no single target (CONVENTIONS.md: empty target when there is none).
-        // Added before the save so it commits atomically with every allocation it describes.
-        _ = await audit.LogAsync(
-            AuditActions.QuotaAllocationReset,
-            string.Empty,
-            string.Empty,
-            new
-            {
-                usersResetCount = resolutions.Count,
-                periodYear = period.Year,
-                periodMonth = period.Month,
-                tierSyncCount,
-            },
-            completionToken);
-
-        try
-        {
-            await dbContext.SaveChangesAsync(completionToken);
-        }
-        catch (DbUpdateException exception) when (resolutions.Any(r => r.IsNew))
-        {
-            // A concurrent reset (or a developer's first /me of the month) inserted some of the rows we
-            // were about to Add. Adopt the winners — re-apply our resolution to their rows — and save
-            // again; a failed SaveChanges leaves every entry (including the audit row) still pending, so
-            // the second save is the same atomic unit. Anything other than a lost race is rethrown.
-            var adopted = await AdoptConcurrentlyCreatedRowsAsync(touched, period, now, completionToken);
-            if (adopted == 0)
-            {
-                throw;
-            }
-
-            logger.LogInformation(exception, "Quota reset for {Period} raced a concurrent writer on {AdoptedCount} allocation(s); adopted the existing rows.", period, adopted);
-            await dbContext.SaveChangesAsync(completionToken);
-        }
-
-        logger.LogInformation(
-            "Quota reset for {Period}: {UsersResetCount} active users, {TierSyncCount} tier syncs.",
-            period,
-            resolutions.Count,
-            tierSyncCount);
-
-        return new QuotaResetResult(resolutions.Count, period.Year, period.Month, now);
-    }
-
-    private static void Stamp(QuotaAllocation allocation, DateTimeOffset now)
-    {
-        allocation.IsHardStopped = false;
-        allocation.ResetDate = now;
-    }
-
-    /// <summary>
-    /// For every allocation in <paramref name="touched"/> that is still <see cref="EntityState.Added"/>
-    /// but whose (user, period) row now exists in the database: detaches ours, copies the resolution
-    /// outputs onto the winner (which keeps its <c>TokensUsed</c>, exactly as a re-resolve would), stamps
-    /// it, and swaps it into <paramref name="touched"/>. Returns how many rows were adopted.
-    /// </summary>
-    private async Task<int> AdoptConcurrentlyCreatedRowsAsync(List<QuotaAllocation> touched, BillingPeriod period, DateTimeOffset now, CancellationToken cancellationToken)
-    {
-        var pendingIds = touched
-            .Where(a => dbContext.Entry(a).State == EntityState.Added)
-            .Select(a => a.UserId)
-            .ToList();
-        if (pendingIds.Count == 0)
-        {
-            return 0;
-        }
-
-        var winners = await dbContext.QuotaAllocations.AsNoTracking()
-            .Where(a => pendingIds.Contains(a.UserId) && a.PeriodYear == period.Year && a.PeriodMonth == period.Month)
-            .ToDictionaryAsync(a => a.UserId, cancellationToken);
-
-        var adopted = 0;
-        for (var i = 0; i < touched.Count; i++)
-        {
-            var ours = touched[i];
-            if (dbContext.Entry(ours).State != EntityState.Added || !winners.TryGetValue(ours.UserId, out var winner))
-            {
-                continue;
-            }
-
-            dbContext.Entry(ours).State = EntityState.Detached;
-
-            winner.AllocatedTokens = ours.AllocatedTokens;
-            winner.ResolvedLevelType = ours.ResolvedLevelType;
-            winner.TierProductId = ours.TierProductId;
-            winner.IsGatewayCapped = ours.IsGatewayCapped;
-            Stamp(winner, now);
-            dbContext.QuotaAllocations.Update(winner);
-
-            touched[i] = winner;
-            adopted++;
-        }
-
-        return adopted;
+        return new QuotaResetResult(outcome.UsersResetCount, outcome.Period.Year, outcome.Period.Month, outcome.ResetDate);
     }
 
     private IQueryable<QuotaAllocation> ForPeriod(BillingPeriod period) =>
