+++ src/FoundryGate.Api/Services/Requests/QuotaRequestService.cs
@@ -2,7 +2,7 @@ using System.Globalization;
 using System.Linq.Expressions;
 using FoundryGate.Api.Services.Audit;
 using FoundryGate.Api.Services.Identity;
-using FoundryGate.Api.Services.Quota;
+using FoundryGate.Core.Quota;
 using FoundryGate.Data;
 using FoundryGate.Data.Entities;
 using FoundryGate.Data.Extensions;
