+++ src/FoundryGate.Api/Services/Users/UserService.cs
@@ -1,11 +1,12 @@
 using System.Globalization;
 using System.Linq.Expressions;
-using FoundryGate.Api.Configuration;
 using FoundryGate.Api.Services.Audit;
 using FoundryGate.Api.Services.Identity;
 using FoundryGate.Api.Services.Keys;
 using FoundryGate.Api.Services.Lifecycle;
 using FoundryGate.Api.Services.Quota;
+using FoundryGate.Core.Configuration;
+using FoundryGate.Core.Quota;
 using FoundryGate.Data;
 using FoundryGate.Data.Entities;
 using FoundryGate.Data.Extensions;
