+++ src/FoundryGate.Core/FoundryGate.Core.csproj
@@ -1,0 +1,14 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <PackageReference Include="Imagile.Framework.Configuration" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\FoundryGate.Data\FoundryGate.Data.csproj" />
+    <ProjectReference Include="..\FoundryGate.Domain\FoundryGate.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
